requirements.txt
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
Flask==2.2.5
idna==3.7
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.0.33
paho-mqtt==2.1.0
protobuf==3.20.3
pycparser==2.21
requests==2.31.0
typing-extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
